requirements.txt
absl-py==2.3.1
adabelief-pytorch==0.2.1
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.1
attrs==25.4.0
batchgenerators==0.25.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.2.1
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.73.0
debugpy==1.8.17
decorator==5.2.1
dicom2nifti==2.6.2
dicomweb-client==0.60.1
diskcache==5.6.3
docker==7.1.0
einops==0.8.1
executing==2.2.1
expiring-dict==1.1.2
expiringdict==1.2.2
fastapi==0.121.2
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
fsspec==2025.10.0
future==1.0.0
gdown==5.2.0
girder-client==3.2.11
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huey==2.5.4
huggingface_hub==1.1.4
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
iopath==0.1.10
ipykernel==7.1.0
ipython==8.13.0
ipython_pygments_lexers==1.1.1
itk==5.4.4.post1
itk-core==5.4.4.post1
itk-filtering==5.4.4.post1
itk-io==5.4.4.post1
itk-numerics==5.4.4.post1
itk-registration==5.4.4.post1
itk-segmentation==5.4.4.post1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==3.2.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
linecache2==1.0.0
lmdb==1.7.5
Mako==1.3.10
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
MedPy==0.5.2
ml_collections==1.1.0
mlflow==3.6.0
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
monai==1.5.1
monailabel==0.8.5
mpmath==1.3.0
msgpack==1.1.2
nest-asyncio==1.6.0
nibabel==5.3.2
ninja==1.13.0
numpy==1.26.4
numpymaxflow==0.0.7
omegaconf==2.3.0
openslide-bin==4.0.0.8
openslide-python==1.4.2
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.3.3
parso==0.8.5
passlib==1.7.4
pillow==12.0.0
platformdirs==4.5.0
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydicom==3.0.1
pydicom-seg==0.4.1
Pygments==2.19.2
PyJWT==2.10.1
pynetdicom==3.0.4
pynrrd==1.1.3
pyparsing==3.2.5
pyrsistent==0.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-gdcm==3.2.2
python-multipart==0.0.20
pytorch-ignite==0.4.11
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
ranger-adabelief==0.1.0
ray==2.51.1
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
retrying==1.4.2
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.6.2
sam2==1.1.0
schedule==1.2.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
simpleitk==2.5.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timeloop==1.0.2
timm==1.0.22
tornado==6.5.2
tqdm==4.67.1
traceback2==1.4.0
traitlets==5.14.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unittest2==1.1.0
urllib3==2.5.0
uvicorn==0.38.0
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.3
zipp==3.23.0
